requirements.txt
pymupdf
langchain
langchain-text-splitters
langchain-community
faiss-cpu
python-dotenv
langchain-google-genai
langchain-google-vertexai
google-auth 
google-cloud-aiplatform
streamlit